+++ resdk/resources/kb/__init__.py
@@ -1,0 +1,17 @@
+""".. Ignore pydocstyle D400.
+
+=========
+Resources
+=========
+
+Resource classes
+================
+
+.. autoclass:: resdk.resources.kb.Feature
+   :members:
+
+"""
+
+from .feature import Feature
+
+__all__ = ('Feature',)