+++ resdk/tests/unit/test_base.py
@@ -17,7 +17,6 @@ class TestBaseResource(unittest.TestCase):
 
     @patch('resdk.resolwe.Resolwe')
     def setUp(self, resolwe_mock):  # pylint: disable=arguments-differ
-        resolwe_mock.configure_mock(api="something")
         self.resolwe_mock = resolwe_mock
 
     @patch('resdk.resources.base.BaseResource', spec=True)
@@ -46,54 +45,53 @@ class TestBaseResource(unittest.TestCase):
         with six.assertRaisesRegex(self, ValueError, message):
             BaseResource.__init__(base_mock, model_data="a", resolwe=self.resolwe_mock)
 
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_id_http_error(self, base_mock, getattr_mock):
+    def test_id_http_error(self, base_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
         base_mock._update_fields.side_effect = slumber.exceptions.HttpNotFoundError
         message = r"ID '\d+' does not exist or you do not have access permission."
         with six.assertRaisesRegex(self, ValueError, message):
             BaseResource.__init__(base_mock, id=1, resolwe=self.resolwe_mock)
 
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_id_ok(self, base_mock, getattr_mock):
+    def test_id_ok(self, base_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
         BaseResource.__init__(base_mock, id=1, resolwe=self.resolwe_mock)
         self.assertEqual(base_mock._update_fields.call_count, 1)
 
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_slug_value_error(self, base_mock, getattr_mock):
-        base_mock.configure_mock(endpoint="this_is_string")
-        getattr_mock.return_value = MagicMock(**{'get.return_value': []})
+    def test_slug_value_error(self, base_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
         message = r"Slug '\w+' does not exist or you do not have access permission."
         with six.assertRaisesRegex(self, ValueError, message):
             BaseResource.__init__(base_mock, slug="a", resolwe=self.resolwe_mock)
         base_mock._update_fields.assert_not_called()
 
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_slug_ok(self, base_mock, getattr_mock):
-        base_mock.configure_mock(endpoint="this_is_string")
-        getattr_mock.return_value = MagicMock(**{'get.return_value': [123]})
+    def test_slug_ok(self, base_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
+        self.resolwe_mock.configure_mock(**{'api.resource_endpoint.get.return_value': [123]})
         BaseResource.__init__(base_mock, slug="a", resolwe=self.resolwe_mock)
         base_mock._update_fields.assert_called_once_with(123)
 
-        getattr_mock.return_value = MagicMock(**{'get.return_value': [{'version': 1},
-                                                                      {'version': 2}]})
+        self.resolwe_mock.configure_mock(
+            **{'api.resource_endpoint.get.return_value': [{'version': 1}, {'version': 2}]}
+        )
+
         BaseResource.__init__(base_mock, slug="a", resolwe=self.resolwe_mock)
         self.assertEqual(base_mock._update_fields.call_count, 2)
 
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_model_data(self, base_mock, getattr_mock):
+    def test_model_data(self, base_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
         mdata = {'a': 1, 'b': 2}
         BaseResource.__init__(base_mock, model_data=mdata, resolwe=self.resolwe_mock)
         base_mock._update_fields.assert_called_once_with(mdata)
 
     @patch('resdk.resources.base.logging')
-    @patch('resdk.resources.base.getattr')
     @patch('resdk.resources.base.BaseResource', spec=True)
-    def test_init_all_ok(self, base_mock, getattr_mock, log_mock):
+    def test_init_all_ok(self, base_mock, log_mock):
+        base_mock.configure_mock(endpoint='resource_endpoint')
         BaseResource.__init__(base_mock, id=1, resolwe=self.resolwe_mock)
         self.assertEqual(log_mock.getLogger.call_count, 1)