+++ resdk/tests/unit/test_resolwe.py
@@ -32,8 +32,9 @@ class TestResolwe(unittest.TestCase):
         Resolwe.__init__(resolwe_mock, 'a', 'b', 'http://some/url')
         self.assertEqual(resauth_mock.call_count, 1)
         self.assertEqual(slumber_mock.API.call_count, 1)
-        # There are four instances of ResolweQuery in init: data, process, sample and collection:
-        self.assertEqual(resolwe_querry_mock.call_count, 5)
+        # There are six instances of ResolweQuery in init: data, process, sample,
+        # presample, collection and feature.
+        self.assertEqual(resolwe_querry_mock.call_count, 6)
         self.assertEqual(log_mock.getLogger.call_count, 1)