+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 -----
 * Check PEP 8 and PEP 257
 * Support delete() method
+* Feature resource and resolwe-update-kb script
 
 Changed
 -------